# [UI Performance] Optimizing recomposition by implementing deferring state read