--- a/app/build.gradle.kts
+++ b/app/build.gradle.kts
@@ -9,6 +9,10 @@ plugins {
 android {
     namespace = "com.example.transpose"
     buildTypes {
+        getByName("release") {
+            signingConfig = signingConfigs.getByName("debug")
+            // 기타 옵션: minifyEnabled 등 원래 release 옵션들 그대로
+        }
         create("benchmark") {
             initWith(buildTypes.getByName("release"))
             signingConfig = signingConfigs.getByName("debug")
@@ -28,6 +32,10 @@ android {
 //        }
 //    }
 
+}
+composeCompiler {
+    reportsDestination = layout.buildDirectory.dir("compose_compiler")
+    metricsDestination = layout.buildDirectory.dir("compose_compiler")
 }
 
 dependencies {