--- a/core/domain/src/main/java/com/example/domain/model/preferences/RepeatMode.kt
+++ b/core/domain/src/main/java/com/example/domain/model/preferences/RepeatMode.kt
@@ -2,11 +2,11 @@ package com.example.domain.model.preferences
 
 
 enum class RepeatMode(val value: Int) {
-    NONE(0),      // 반복 없음
-    ALL(1),       // 전체 반복
-    ONE(2);       // 한 곡 반복
+    OFF(0),
+    ALL(1),
+    ONE(2);
 
     companion object {
-        fun fromValue(value: Int) = entries.find { it.value == value } ?: NONE
+        fun fromValue(value: Int) = entries.find { it.value == value } ?: OFF
     }
 }