--- a/core/domain/src/main/java/com/example/domain/repository/VideoRepository.kt
+++ b/core/domain/src/main/java/com/example/domain/repository/VideoRepository.kt
@@ -1,7 +1,9 @@
 package com.example.domain.repository
 
 import com.example.domain.model.youtube.video_detail.VideoDetail
+import kotlinx.coroutines.flow.StateFlow
 
 interface VideoRepository {
-    suspend fun fetchVideoDetail(videoId: String): Result<VideoDetail>
+    val currentVideoDetail: StateFlow<VideoDetail?>
+    suspend fun fetchVideoDetail(videoId: String)
 }