--- a/core/domain/src/main/java/com/example/domain/repository/VideoRepository.kt
+++ b/core/domain/src/main/java/com/example/domain/repository/VideoRepository.kt
@@ -4,6 +4,5 @@ import com.example.domain.model.youtube.video_detail.VideoDetail
 import kotlinx.coroutines.flow.StateFlow
 
 interface VideoRepository {
-    val currentVideoDetail: StateFlow<VideoDetail?>
-    suspend fun fetchVideoDetail(videoId: String)
+    suspend fun fetchVideoDetail(videoId: String): Result<VideoDetail>
 }