--- a/core/ui/src/main/java/com/example/ui/screen/playlist_info/PlaylistInfoViewModel.kt
+++ b/core/ui/src/main/java/com/example/ui/screen/playlist_info/PlaylistInfoViewModel.kt
@@ -97,7 +97,7 @@ class PlaylistInfoViewModel @Inject constructor(
         }
 
 
-    fun onMediaClicked(
+    fun playPlaylist(
         playlistItems: List<Video>,
         clickedIndex: Int
     ) {