--- a/core/ui/src/main/java/com/example/ui/screen/search_result/SearchResultScreen.kt
+++ b/core/ui/src/main/java/com/example/ui/screen/search_result/SearchResultScreen.kt
@@ -90,7 +90,7 @@ fun SearchResultScreen(
                                     coroutineScope.launch {
                                         bottomSheetState.expand()
                                     }
-                                    searchResultViewModel.onMediaClicked(
+                                    searchResultViewModel.playSingleVideo(
                                         item.video
                                     )
                                 },