--- a/feature/home/src/main/java/com/example/home/home_playlist/HomePlaylistViewModel.kt
+++ b/feature/home/src/main/java/com/example/home/home_playlist/HomePlaylistViewModel.kt
@@ -6,6 +6,7 @@ import com.example.domain.constants.MusicCategoryConstants
 import com.example.domain.model.youtube.playlist.Playlist
 import com.example.domain.repository.PlaylistRepository
 import com.example.media.manager.MediaPlaybackManager
+import com.example.media.state_holder.NowPlayingStateHolder
 import com.example.ui.common.UiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
@@ -18,7 +19,7 @@ import javax.inject.Inject
 @HiltViewModel
 class HomePlaylistViewModel @Inject constructor(
     private val playlistRepository: PlaylistRepository,
-    private val mediaPlaybackManager: MediaPlaybackManager
+    private val nowPlayingStateHolder: NowPlayingStateHolder,
 ) : ViewModel() {
 
     private val _nationalPlaylistDataState =
@@ -42,7 +43,7 @@ class HomePlaylistViewModel @Inject constructor(
     }
 
     fun setCurrentPlaylistInfo(playlist: Playlist) {
-        mediaPlaybackManager.setCurrentPlaylistInfo(playlist)
+        nowPlayingStateHolder.setCurrentPlaylistInfo(playlist)
     }