--- a/feature/main/src/main/java/com/example/main/MainScreen.kt
+++ b/feature/main/src/main/java/com/example/main/MainScreen.kt
@@ -169,6 +169,7 @@ fun MainScreen(
                         updateSearchBarState(SearchBarState.CLOSED)
                     },
                     mainViewModel = mainViewModel,
+                    updateSearchBarState = updateSearchBarState,
                     searchBarState = searchBarState,
                     scrollBehavior = scrollBehavior,
                 )