## Changes

- Created `NowPlayingStateHolder` to separate current media content management from media control logic.
- Hoisted the BottomSheet normalization logic (offset calculation, etc.), making the logic easier to manage and reuse.
- Used deferred state read and lambda-based modifiers to help skip unnecessary composition phases, making the UI more efficient.


## Recomposition Optimization Results

<table width="100%">
  <tr>
    <td align="center"><b>Before</b></td>
    <td align="center"><b>After</b></td>
  </tr>
  <tr>
    <td style="width:50%">
      <img src="https://github.com/user-attachments/assets/cf2c4fce-2b24-438e-9211-56ae564b6e43" style="width:100%; height:auto;"/>
    </td>
    <td style="width:50%">
      <img src="https://github.com/user-attachments/assets/14760690-9ac2-465f-90a1-04463d936826" style="width:100%; height:auto;"/>
    </td>
  </tr>
</table>

The GIFs show the difference before and after recomposition optimization.

- **Before:** The entire layout frequently recomposes whenever the state changes.

- **After:** Only the necessary parts recompose, resulting in a much lower recomposition count.

## Benchmark Result Comparison

### Macrobenchmark Results (Before)

| Metric                    | Min   | Median | Max   | P50   | P90   | P95   | P99   |
|---------------------------|-------|--------|-------|-------|-------|-------|-------|
| frameCount                | 135.0 | 157.0  | 218.0 |   –   |   –   |   –   |   –   |
| timeToInitialDisplayMs    | 523.3 | 556.7  | 710.4 |   –   |   –   |   –   |   –   |
| frameDurationCpuMs        |   –   |   –    |   –   | 6.9   | 22.1  | 27.2  | 67.1  |
| frameOverrunMs            |   –   |   –    |   –   | -3.0  | 13.8  | 33.6  | 283.9 |


---


### Macrobenchmark Results (After)


| Metric                   | Min    | Median  | Max    | P50   | P90   | P95   | P99   |
|--------------------------|--------|---------|--------|-------|-------|-------|-------|
| frameCount               | 285.0  | 300.0   | 400.0  |   –   |   –   |   –   |   –   |
| timeToInitialDisplayMs   | 527.7  | 563.1   | 614.6  |   –   |   –   |   –   |   –   |
| frameDurationCpuMs       |   –    |   –     |   –    | 9.2   | 25.7  | 28.9  | 54.1  |
| frameOverrunMs           |   –    |   –     |   –    | -1.8  | 15.2  | 17.0  | 82.7  |


### Perfetto Frame Timing Comparison

| Metric                | Before         | After          |
|-----------------------|---------------|----------------|
| First Frame Duration  | 304 ms 8 μs  |   259 ms 12 us  |
| Measure Phase Time    | 112 ms 744 μs  | 114 ms 467 μs  |

- [Before Trace (Google Drive)](https://drive.google.com/file/d/1ijZo13xuy--aILhohFmVXlxLX1I_PVWa/view?usp=sharing)
- [After Trace (Google Drive)](https://drive.google.com/file/d/1APFazmYq1nPJpX-MWeJEZwZkZwYxJPN-/view?usp=sharing)


## Key Findings

- I found out that using a state setter like `val (bottomSheetOffset, updateBottomSheetOffset) = remember { mutableStateOf(0f) }` causes composable that has the state to recompose every time the state changes, regardless of where the state is actually read. By switching to a simple `var` for the state, I was able to prevent unnecessary recompositions of the parent.